--- iidm/iidm-mergingview/src/main/java/com/powsybl/iidm/mergingview/VoltageLevelBusViewAdapter.java
@@ -1,46 +1,0 @@
-/**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.powsybl.iidm.mergingview;
-
-import com.powsybl.iidm.network.Bus;
-import com.powsybl.iidm.network.VoltageLevel;
-import com.powsybl.iidm.network.VoltageLevel.BusView;
-
-import java.util.stream.Stream;
-
-/**
- * @author Thomas Adam <tadam at silicom.fr>
- */
-public class VoltageLevelBusViewAdapter extends AbstractAdapter<VoltageLevel.BusView> implements VoltageLevel.BusView {
-
-    VoltageLevelBusViewAdapter(final BusView delegate, final MergingViewIndex index) {
-        super(delegate, index);
-    }
-
-    // -------------------------------
-    // Not implemented methods -------
-    // -------------------------------
-    @Override
-    public Iterable<Bus> getBuses() {
-        throw MergingView.NOT_IMPLEMENTED_EXCEPTION;
-    }
-
-    @Override
-    public Stream<Bus> getBusStream() {
-        throw MergingView.NOT_IMPLEMENTED_EXCEPTION;
-    }
-
-    @Override
-    public Bus getBus(final String id) {
-        throw MergingView.NOT_IMPLEMENTED_EXCEPTION;
-    }
-
-    @Override
-    public Bus getMergedBus(final String configuredBusId) {
-        throw MergingView.NOT_IMPLEMENTED_EXCEPTION;
-    }
-}
